first pass at comparison of desired and actual state
instead of just blindly pushing data to APIC we would prefer to
determine it's current state and only push any changes necessary to
alter it's state to that declared by the user. Currently these changes
could include adding deleting or modifying nodes. This commit is a naive
initial implementation of this.